Many Changes. See description.
* Made changes to the particle field rendering, thus being able to
provide 2 config options: Number of rendered particle planes. The more
the merrier :D. And an option to turn of the particlefield entirely and
just render a static image. Closes #53
* Added added color markers for each direction to the DP, so you can see
which side you are working on the inside without hitting F3 to check
* Since those markers are not nice to look at from the inside of a
pocket, they are only visible when the player is holding a Nether
Crystal in his hand